requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==6.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cmdstanpy==1.2.5
contourpy==1.3.2
curl_cffi==0.11.4
cycler==0.12.1
Cython==3.1.2
distro==1.9.0
exceptiongroup==1.3.0
fonttools==4.58.4
frozendict==2.4.6
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
holidays==0.75
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
multitasking==0.0.11
narwhals==1.44.0
numpy==1.24.4
openai==1.92.0
packaging==25.0
pandas==2.3.0
patsy==1.0.1
peewee==3.18.1
pillow==11.2.1
platformdirs==4.3.8
pmdarima==2.0.4
prophet==1.1.7
protobuf==6.31.1
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.25.1
scikit-learn==1.7.0
scipy==1.15.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
stanio==0.5.1
statsmodels==0.14.4
streamlit==1.46.1
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
websockets==15.0.1
xgboost==3.0.2
yfinance==0.2.63
